Modified ballDrop: Still IDE issue with Chapter4Notes

+++ ballDrop/ballDrop/ballDrop.cpp
@@ -4,13 +4,13 @@
 
 int main()
 {
-    int height = collectHeight();                               // Collect the Height of the building: Look at collectData.h                         
-    int seconds = collectSeconds();                             // Collect the period of the ball drop: Look at collectData.h
-    std::cout << "If the building's height is " << height       // Printing out the following results
+    double height = collectHeight();                               // Collect the Height of the building: Look at collectData.h                         
+    double seconds = collectSeconds();                             // Collect the period of the ball drop: Look at collectData.h
+    std::cout << "If the building's height is " << height          // Printing out the following results
         << "m, then the ball distance will be the following " 
         << seconds << " seconds." << '\n';
     
-    condition(seconds, height);                                 // The actual math: Look at condition.h
+    condition(seconds, height);                                     // The actual math: Look at condition.h
 
  
 }
