+++ ballDrop/ballDrop/collectData.h
@@ -2,17 +2,17 @@
 #define COLLECTDATA_H
 #include <iostream>
 
-int collectSeconds() {																		// Collect Data that how long the user want to see 
+double collectSeconds() {																	// Collect Data that how long the user want to see 
 																							// the ball dropping form the building
-	int x{};
+	double x{};
 	std::cout << "Enter the seconds that you want to see the ball dropping down: ";
 	std::cin >> x;
 	std::cout << '\n';
 	return x;
 }
 
-int collectHeight() {																		// Collect Data that how tall the building is
-	int x{};
+double collectHeight() {																	// Collect Data that how tall the building is
+	double x{};
 	std::cout << "Enter the building height in meters: ";
 	std::cin >> x;
 	std::cout << '\n';
